feat: Character Repository 구현
- Core Data 여부 판단 하고 없을 때 네트워크 요청하도록 구현

+++ Tekken8 Frame Data/Tekken8 Frame Data/Character/ViewModel/CharacterListViewModel.swift
@@ -10,7 +10,7 @@ import UIKit
 
 @MainActor
 final class CharacterListViewModel: ObservableObject {
-    @Published private(set) var characterImages: [Int: UIImage] = [:]
+    @Published private(set) var characterImages: [Int64: UIImage] = [:]
     @Published private(set) var filteredCharacters: [Character] = []
     private(set) var characters: [Character] = []
     private let repository: CharacterRepository
